backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Mail==0.9.1

# Database
pymongo==4.6.1
redis==5.0.1
qdrant-client==1.7.3

# AI/ML with Google Gemini
google-generativeai==0.3.2
sentence-transformers==2.3.1
spacy==3.7.2
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2

# Task Queue
celery==5.3.6
redis

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
pytest==7.4.3
black==23.12.1
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
email-validator==2.1.0.post1
